Inline the get_mesh calls
This will be better for memory because the `Mesh` objects are created right before the `draw` call, inside the parameter with a reference operator attached to it, meaning each `Mesh` is dropped right after the call and before the other one is instantiated. Because this operation will take the same about of time total, it will help lower memory usage because memory is freed right after use instead of at the end of the function call, specifically after the OS yield. Although Rust probably already does this through non-lexical lifetimes, doing it idiomatically might help.
Also bump the version because I forgot to earlier